start working on ModelServing examples.
